+++ ElectronNET.API/Entities/WebPreferences.cs
@@ -185,7 +185,8 @@ namespace ElectronNET.API.Entities
         /// Context' entry in the combo box at the top of the Console tab. This option is
         /// currently experimental and may change or be removed in future Electron releases.
         /// </summary>
-        public bool ContextIsolation { get; set; }
+        [DefaultValue(true)]
+        public bool ContextIsolation { get; set; } = true;
 
         /// <summary>
         /// Whether to use native window.open(). Defaults to false. This option is currently experimental.
@@ -203,6 +204,7 @@ namespace ElectronNET.API.Entities
         /// <value>
         ///   <c>true</c> if [webview tag]; otherwise, <c>false</c>.
         /// </value>
-        public bool WebviewTag { get; set; }
+        [DefaultValue(false)]
+        public bool WebviewTag { get; set; } = false;
     }
 }