+++ ElectronNET.Host/api/app.ts
@@ -268,10 +268,6 @@ export = (socket: SocketIO.Socket, app: Electron.App) => {
         app.commandLine.appendArgument(value);
     });
 
-    socket.on('appEnableMixedSandbox', () => {
-        app.enableMixedSandbox();
-    });
-
     socket.on('appDockBounce', (type) => {
         const id = app.dock.bounce(type);
         electronSocket.emit('appDockBounceCompleted', id);